+++ packages/highlight/postcss.config.js
@@ -1,6 +1,7 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
 module.exports = {
   plugins: [
-    require('tailwindcss')(), 
-    require('autoprefixer')()
+    require("tailwindcss")(),
+    require("autoprefixer")(),
   ],
-}
+};