+++ packages/highlight/src/utils/getComponents.ts
@@ -1,0 +1,5 @@
+import components from "prismjs/components.json";
+
+export function getComponents() {
+  return components;
+}