+++ packages/storybook/src/components/base/List/index.ts
@@ -1,0 +1,15 @@
+import { tf } from "tailwind-factory";
+
+export const List = tf("ul", `
+  pl-6
+  lg:pl-10
+  my-2
+`, {
+  variants: {},
+  defaultVariants: {}
+});
+
+export const ListItem = tf("li", ``, {
+  variants: {},
+  defaultVariants: {}
+});